update des comparaison Integer Integer en int Integer

+++ src/main/java/fr/eni/enchere/controller/ServletDetailVente.java
@@ -87,7 +87,8 @@ public class ServletDetailVente extends HttpServlet {
 					request.setAttribute("encheres", encheres);
 					System.out.println(av.toString());
 					//Si l'utilisateur consulte une de ses ventes
-					if ((Integer)request.getSession().getAttribute("noUtilisateur") == av.getNoUtilisateur()) {
+					
+					if ((Integer)request.getSession().getAttribute("noUtilisateur") == (int)av.getNoUtilisateur()) {
 						RequestDispatcher rd = request.getRequestDispatcher("/WEB-INF/jsp/detailMaVente.jsp");
 						rd.forward(request, response);
 					} else {
